+++ src/util/ALCheckDocumentation.ts
@@ -1,4 +1,4 @@
-import { Diagnostic, DiagnosticCollection, languages, Range, TextDocument } from 'vscode';
+import { Diagnostic, Range, TextDocument } from 'vscode';
 import { ALXmlDocDiagnosticCode, ALXmlDocDiagnosticPrefix } from '../types';
 import { ALObject } from '../types/ALObject';
 import { ALObjectType } from '../types/ALObjectType';
@@ -7,72 +7,40 @@ import { XMLDocumentationExistType } from '../types/XMLDocumentationExistType';
 import { ALSyntaxUtil } from './ALSyntaxUtil';
 import { ALDocCommentUtil } from './ALDocCommentUtil';
 import { Configuration } from './Configuration';
-import { StringUtil } from './StringUtil';
-import * as fs from 'fs';
+import { StringUtil } from './string/StringUtil';
 import { ALProcedureType } from '../types/ALProcedureType';
+import { ALCodeunitType } from '../types/ALCodeunitType';
+import { ALDiagnosticCollection } from './ALDiagnosticCollection';
 
-export class ALCheckDocumentation {
-    /**
-     * Collection to store all gathered diagnostics.
-     */
-    private static diagCollection: DiagnosticCollection = languages.createDiagnosticCollection(ALXmlDocDiagnosticPrefix);
-    
+export class ALCheckDocumentation {    
     /**
      * Gathered diagnostics.
      */
-    private static diags: Diagnostic[] = [];
+    private diags: Diagnostic[] = [];
 
     /**
      * Actual AL Source Code file.
      */
-    private static document: any;
+    private document: any;
 
     /**
      * AL Object.
      */
-    private static alObject: ALObject | null = null;
+    private alObject: ALObject;
 
-    /**
-     * Check documentation for passed TextDocument.
-     * @param document {TextDocument}
-     */
-    public static CheckDocumentationForDocument(document: TextDocument) {
-        if (document.languageId !== 'al') {
-            return;
-        }
+    public constructor(document: TextDocument, alObject: ALObject) {
         this.document = document;
-
-        this.CheckDocumentation();
-    }
-
-    /**
-     * Check documentation for passed AL Object.
-     * @param alObject {ALObject}
-     */
-    public static CheckDocumentationForALObject(alObject: ALObject, document: TextDocument | undefined = undefined) {
-        if (document === undefined) {
-            this.document = Object.assign({});
-            this.document.getText = () => fs.readFileSync(`${alObject.Path}/${alObject.FileName}`, 'utf8');
-            this.document.fileName = `${alObject.Path}/${alObject.FileName}`;
-            this.document.uri = alObject.Uri;
-        } else {
-            this.document = document;
-
-            if (document.languageId !== 'al') {
-                return;
-            }
-        }
-
         this.alObject = alObject;
-
-        this.CheckDocumentation();
+        
+        // clear all diagnostics
+        this.diags = [];
+        this.UpdateDiagnosticCollection();
     }
 
     /**
      * General documentation check procedure.
      */
-    private static CheckDocumentation() {
-        this.Initialize();
+    public CheckDocumentation() {
         
         if ((!Configuration.ProcedureDocumentationCheckIsEnabled(this.document.uri)) && (!Configuration.ObjectDocumentationCheckIsEnabled(this.document.uri))) {
             return;
@@ -85,6 +53,12 @@ export class ALCheckDocumentation {
         if (Configuration.ObjectDocumentationCheckIsEnabled(this.document.uri)) {
             this.AnalyzeObjectDocumentation(this.alObject);
         }
+
+        // do not check any kind of procedures in 'Test' codeunits.
+        if ((this.alObject.Subtype === ALCodeunitType.Test) && (!Configuration.IsDocumentationMandatoryForTest())) {
+            return;
+        }
+
         if (Configuration.ProcedureDocumentationCheckIsEnabled(this.document.uri)) {
             this.alObject.Procedures?.forEach(alProcedure => {
                 this.AnalyzeProcedureDocumentation(this.alObject, alProcedure);
@@ -95,46 +69,38 @@ export class ALCheckDocumentation {
         
         this.UpdateDiagnosticCollection();
     }
-
-    /**
-     * Initialize documentation check and clear previously reported diagnostics.
-     */
-    private static Initialize() {
-        // clear all diagnostics
-        this.diags = [];
-        this.UpdateDiagnosticCollection();
-
-        if (this.alObject === null) {
-            this.alObject = ALSyntaxUtil.GetALObject(this.document);
-        }
-    }
     
     /**
      * Analyze source code for unnecessary XML documentations.
      * @param alObject ALObject
      * @param document TextDocument
      */
-    private static AnalyzeUnnecessaryDocumentation(alObject: ALObject | null, document: TextDocument) {
+    private AnalyzeUnnecessaryDocumentation(alObject: ALObject | null, document: TextDocument) {
         if (alObject === null) {
             return;
         }
         let i = 0;
         try {
             let codeLines: string[] = ALSyntaxUtil.SplitALCodeToLines(document.getText());
-            let xmlDocumentation: string = '';
+            let xmlDocumentation: string ='';
             for(i = 0; i < codeLines.length; i++) {
                 let line: string = codeLines[i];
                 if (!line.trim().startsWith('///')) {
-                    if (xmlDocumentation !== '') {
-                        let alProcedure: ALProcedure | undefined = alObject.Procedures?.find(alProcedure => (i <= alProcedure.LineNo));
-                        if (alProcedure === undefined) {
-                            console.debug(`Could not find AL Procedure for XML documentation found in ${alObject.FileName} line ${i}.`);
-                            continue;
+                    if (ALSyntaxUtil.IsObjectDefinition(line)) {
+                        xmlDocumentation = '';
+                        continue;
+                    } else {
+                        if (xmlDocumentation !== '') {
+                            let alProcedure: ALProcedure | undefined = alObject.Procedures?.find(alProcedure => (i <= alProcedure.LineNo));
+                            if (alProcedure === undefined) {
+                                console.debug(`Could not find AL Procedure for XML documentation found in ${alObject.FileName} line ${i}.`);
+                                continue;
+                            }
+                            this.GetUnnecessaryProcedureDocumentationDiagnostics(codeLines, i, xmlDocumentation, alObject, alProcedure);
                         }
-                        this.GetUnnecessaryProcedureDocumentationDiagnostics(codeLines, i, xmlDocumentation, alObject, alProcedure);
+                        xmlDocumentation = '';
+                        continue;
                     }
-                    xmlDocumentation = '';
-                    continue;
                 }
 
                 xmlDocumentation += line.trim().replace('///','');
@@ -152,7 +118,7 @@ export class ALCheckDocumentation {
      * @param alObject ALObject
      * @param alProcedure ALProcedure
      */
-    private static GetUnnecessaryProcedureDocumentationDiagnostics(codeLines: string[], currentLineNo: number, xmlDocumentation: string, alObject: ALObject, alProcedure: ALProcedure) {
+    private GetUnnecessaryProcedureDocumentationDiagnostics(codeLines: string[], currentLineNo: number, xmlDocumentation: string, alObject: ALObject, alProcedure: ALProcedure) {
         // convert to JSON to make it more accessible 
         let jsonDocumentation = ALDocCommentUtil.GetJsonFromXmlDocumentation(xmlDocumentation);
         if (!jsonDocumentation.param) {
@@ -200,7 +166,11 @@ export class ALCheckDocumentation {
      * @param param Documented parameter
      * @param alProcedure ALProcedure
      */
-    private static GetUnnecessaryParameterDocumentationDiagnostics(unnecessaryParameters: Array<string>, param: { value: string, attr: { name: string }}, alProcedure: ALProcedure) {
+    private GetUnnecessaryParameterDocumentationDiagnostics(unnecessaryParameters: Array<string>, param: { value: string, attr: { name: string }}, alProcedure: ALProcedure) {
+        if (!param) {
+            return;
+        }
+        
         if (alProcedure.Parameters.find(alParameter => (alParameter.Name === param.attr.name)) === undefined) {
             unnecessaryParameters.push(param.attr.name);
         }
@@ -211,12 +181,12 @@ export class ALCheckDocumentation {
      * @param alObject {ALObject}
      * @param alProcedure {ALProcedure}
      */
-    private static AnalyzeProcedureDocumentation(alObject: ALObject | null, alProcedure: ALProcedure) {
+    private AnalyzeProcedureDocumentation(alObject: ALObject | null, alProcedure: ALProcedure) {
         if (alObject === null) {
             return;
         }
 
-        if (!Configuration.IsDocumentationMandatory(alProcedure.Type, alProcedure.Subtype, alProcedure.Access, alObject.Uri)) {
+        if (!Configuration.IsProcedureDocumentationMandatory(alObject.Access, alProcedure.Type, alProcedure.Subtype, alProcedure.Access, alObject.Uri)) {
             return;
         }
 
@@ -316,8 +286,12 @@ export class ALCheckDocumentation {
      * Analyse documentation of the given object.
      * @param alObject 
      */
-    private static AnalyzeObjectDocumentation(alObject: ALObject) {
+    private AnalyzeObjectDocumentation(alObject: ALObject) {
         if (alObject === null) {
+            return;
+        }
+
+        if (!Configuration.IsDocumentationMandatoryForAccessLevel(alObject.Access)) {
             return;
         }
         
@@ -336,7 +310,7 @@ export class ALCheckDocumentation {
      * Checks whether any XML documentation is present for AL Procedure or not.
      * @param alProcedure ALProcedure to search for any XML documentation present
      */
-    private static IsMissingXmlDocumentation(alProcedure: ALProcedure): boolean {
+    private IsMissingXmlDocumentation(alProcedure: ALProcedure): boolean {
         return (alProcedure.XmlDocumentation.Exists === XMLDocumentationExistType.No) && 
             ((alProcedure.Return === undefined) || (alProcedure.Return?.XmlDocumentation.Exists === XMLDocumentationExistType.No)) && 
             (alProcedure.Parameters?.find(alProcedure => (alProcedure.XmlDocumentation.Exists === XMLDocumentationExistType.Yes)) === undefined);
@@ -346,14 +320,14 @@ export class ALCheckDocumentation {
      * Returns diagnostics string.
      * @param diagnosticCode {ALXmlDocDiagnosticCode}
      */
-    private static GetDiagnosticCode(diagnosticCode: ALXmlDocDiagnosticCode): any {
+    private GetDiagnosticCode(diagnosticCode: ALXmlDocDiagnosticCode): any {
         return diagnosticCode.toString();
     }
 
     /**
      * Returns true if actual diagnostic code is representing a missing documentation.
      */
-    private static IsMissingDocumentationDiag(element: { diagnosticCode: ALXmlDocDiagnosticCode; }, index: any, array: any) {
+    private IsMissingDocumentationDiag(element: { diagnosticCode: ALXmlDocDiagnosticCode; }, index: any, array: any) {
         return (
             (element.diagnosticCode === ALXmlDocDiagnosticCode.XmlDocumentationMissing) || 
             (element.diagnosticCode === ALXmlDocDiagnosticCode.SummaryMissing) ||
@@ -364,20 +338,20 @@ export class ALCheckDocumentation {
     /**
      * Returns true if actual diagnostic code is representing a unnecessary documentation.
      */
-    private static IsUnnecessaryDocumentationDiag(element: { diagnosticCode: ALXmlDocDiagnosticCode; }, index: any, array: any) {
+    private IsUnnecessaryDocumentationDiag(element: { diagnosticCode: ALXmlDocDiagnosticCode; }, index: any, array: any) {
         return ((element.diagnosticCode === ALXmlDocDiagnosticCode.ParameterUnnecessary));
     }
 
     /**
      * Update DiagnosticCollection to present them to the user.
      */
-    private static UpdateDiagnosticCollection() {
+    private UpdateDiagnosticCollection() {
         if (this.diags === []) {
-            this.diagCollection.clear();
+            ALDiagnosticCollection.Diagnostics.clear();
             return;
         }
         
-        this.diagCollection.set(this.document.uri, this.diags);
+        ALDiagnosticCollection.Diagnostics.set(this.document.uri, this.diags);
     }
 
 }