+++ src/util/ALLangServerProxy.ts
@@ -213,7 +213,7 @@ export class ALLangServerProxy {
             document.fileName = (alDefinition.uri.scheme === 'al-preview') ? '__symbol__' : alDefinition.uri.fsPath;
 
             let result: { ALObject: ALObject | null, Position: Position} = {
-                ALObject: ALSyntaxUtil.GetALObject(document),
+                ALObject: await ALSyntaxUtil.GetALObject(document),
                 Position: alDefinition.range.end
             };
 
@@ -221,7 +221,7 @@ export class ALLangServerProxy {
         }
         catch(ex) 
         {
-            console.error(`[GetALObjectFromDefinition] - ${ex} Please report this Please report this error at https://github.com/365businessdev/vscode-alxmldocumentation/issues`);
+            console.error(`[GetALObjectFromDefinition] - ${ex} Please report this error at https://github.com/365businessdev/vscode-alxmldocumentation/issues`);
             return undefined;
         }